+++ apps/docs/CHANGELOG.md
@@ -1,9 +1,9 @@
-# @stella-ui/docs
+# @stella-ds/docs
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [747cbf6]
-  - @stella-ui/react@0.2.0
-  - @stella-ui/theme@0.2.0
+  - @stella-ds/react@0.2.0
+  - @stella-ds/theme@0.2.0