+++ packages/theme/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @stella-ui/theme
+# @stella-ds/theme
 
 ## 0.2.0
 
 ### Minor Changes
 
-- 747cbf6: Initial release of @stella-ui/react and @stella-ui/theme
+- 747cbf6: Initial release of @stella-ds/react and @stella-ds/theme